remove JsonPathInst, instead move `find` functionality to `JsonPath`. no longer export all macros, add comments to the remaining ones. Implement `FromStr` and `TryFrom<&str` for JsonPath (as the internet is not sure which is better) fix various clippy things.
